// challenge-2/frontend/src/shared/types/error.types.ts
export enum EErrorTypes {
  UNKNOWN_ERROR = 'UNKNOWN_ERROR',
  NON_ETHEREUM_BROWSER = 'NON_ETHEREUM_BROWSER',
  WEB3_ACCESS_REJECTED = 'WEB3_ACCESS_REJECTED',
  WEB3_NOT_INITIALISED = 'WEB3_NOT_INITIALISED',
  WEB3_NO_ACCOUNT_FOUND = 'WEB3_NO_ACCOUNT_FOUND',
  WEB3_NO_PROVIDER_FOUND = 'WEB3_NO_PROVIDER_FOUND',
  WEB3_USER_DECLINED_SIGNING = 'WEB3_USER_DECLINED_SIGNING',
  ABI_INVALID = 'ABI_INVALID',
  API_ERROR = 'API_ERROR'
}